% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_assay.R
\name{parse_assay}
\alias{parse_assay}
\title{Parse an assay in an exported BCS XP ASCII file}
\usage{
parse_assay(chunk, include_subassays = FALSE)
}
\arguments{
\item{chunk}{A vector of lines from an assay chunk}

\item{include_subassays}{Whether to include the subassays. Adds a subassay list column to the output. Defaults to FALSE}
}
\description{
This function parses an assay chunk from a BCS XP ASCII file. It is called from read_bcsxp()
and probably doesn't need to be called on it's own. Returns a list of assay parameters.
}
\examples{
parse_assay(readLines(path = "data/S201911271.BCSXp")[3:7])
}
\keyword{BCS}
\keyword{XP}
\keyword{analyzer}
\keyword{coagulation}
