% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_assays.R
\name{read_assays}
\alias{read_assays}
\title{Read a BCS XP S-file file}
\usage{
read_assays(chunks, include_subassays = FALSE, header)
}
\arguments{
\item{chunks}{A list of text from BCS XP ASCII export files, broken up into chunks}

\item{include_subassays}{Adds a subassay list column to the output. Defaults to FALSE}

\item{header}{Manually pass along the file header so it's details can be returned}
}
\description{
This function reads an ASCII-formated sample export file ("S-files") from a BCS XP coagulation analyzer
and returns the assay results into a tidy tibble.
}
\keyword{BCS}
\keyword{XP}
\keyword{analyzer}
\keyword{coagulation}
