% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bcsxp.R
\name{read_bcsxp}
\alias{read_bcsxp}
\title{Read a BCS XP data file}
\usage{
read_bcsxp(path, filetype = "guess", include_subassays = FALSE)
}
\arguments{
\item{path}{Path to the *.BCSXp file}

\item{filetype}{The type of file to be read. One of c("S", "C", or "guess").}

\item{include_subassays}{Adds a subassay list column to the output. Defaults to FALSE}
}
\description{
This function reads an ASCII-formated sample export file ("S-files") from a BCS XP coagulation analyzer
and returns the assay results into a tidy tibble.
}
\examples{
read_bcsxp(path = "data/S201911271.BCSXp", include_subassays = TRUE)
}
\keyword{BCS}
\keyword{XP}
\keyword{analyzer}
\keyword{coagulation}
