% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_curves.R
\name{read_curves}
\alias{read_curves}
\title{Read a BCS XP C-file file}
\usage{
read_curves(chunks, header)
}
\arguments{
\item{chunks}{A list of text from BCS XP ASCII export files, broken up into chunks}

\item{header}{Manually pass along the file header so it's details can be returned}
}
\description{
This function reads an ASCII-formated calibration export file ("C-files") from a BCS XP coagulation analyzer
and returns the calibration data results into a tidy tibble.
}
\keyword{BCS}
\keyword{XP}
\keyword{analyzer}
\keyword{coagulation}
