% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rawfile.R
\name{read_rawfile}
\alias{read_rawfile}
\title{Read a BCS XP R-file file}
\usage{
read_rawfile(chunks, header)
}
\arguments{
\item{chunks}{A list of text from BCS XP ASCII export files, broken up into chunks}

\item{header}{Manually pass along the file header so it's details can be returned}
}
\description{
This function reads an ASCII-formated raw data file ("R-files") from a BCS XP coagulation analyzer
and returns the details on the assay, as well as a list column that contains absorbance vs. time curves
}
\keyword{BCS}
\keyword{XP}
\keyword{analyzer}
\keyword{coagulation}
