% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_chunkstart.R
\name{is_chunkstart}
\alias{is_chunkstart}
\title{Check if a character vector is the start of an assay chunk}
\usage{
is_chunkstart(x)
}
\arguments{
\item{x}{A character string}
}
\description{
This function checks if a given character vector "looks" like the start of an assay chunk from a BCS XP ASCII export
First it checks if the string is tab separated with a length of 2 or greater
If that's true, it checks if the second tab is an "S" (i.e., a sample) or a "C" (i.e., a control).
Returns TRUE or FALSE
}
\keyword{BCS}
\keyword{XP}
\keyword{analyzer}
\keyword{coagulation}
